refactor: extract error parsing and status helpers into core modules
Moves `parse_streamrip_error` into `src/core/errors.py` and
extracts the status-bar color table + markup builder into
`src/core/status.py`. `AurynApp._set_status` becomes a one-liner
that delegates to `build_status_markup`. No logic changes.
